@@ -1,6 +1,7 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-
+import { createRoot } from 'react-dom/client';
 import App from './App';
+const container = document.getElementById('root');
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const root = createRoot(container)
+
+root.render( < App / > );
