@@ -13,7 +13,7 @@ const Cards = ({data: {confirmed, recovered, deaths, lastUpdate }}) =>{
   }
   return(
     <div className={styles.container}>
-      <Grid container spacing={3} justify="center">
+      <Grid container spacing={3} justifyContent="center">
         <Grid item component={Card} xs={12} md={3} className={cx(styles.card, styles.infected)}>
           <CardContent>
             <Typography color="textSecondary" gutterBottom>Infected</Typography>
