@@ -1,8 +1,11 @@
 import React, {useState, useEffect} from 'react';
 import { fetchDailyData } from '../../api';
+import { Chart as ChartJS, registerables } from 'chart.js';
 import { Line, Bar } from 'react-chartjs-2';
-
 import styles from './Chart.module.css'; 
+ChartJS.register(...registerables);
+
+
 
 const Chart = ({data:{ confirmed, deaths, recovered}, country}) =>{
   const [dailyData, setDailyData] = useState([]);
